refactor: Change convention from Requests to Api
