+++ playground/vanilla/app/javascript/api/UserPreferencesApi.ts
