+++ playground/vanilla/app/javascript/api/VideoClipsApi.ts
