+++ playground/vanilla/app/javascript/entrypoints/application.ts
@@ -6,13 +6,11 @@ import Videos from '~/Videos.vue'
 
 // Example: Combine all exported routes in a single object.
 import api from '~/ApiHelpers'
-import * as ApiComposables from '~/composables/api'
-
-const app = createApp(Videos)
 
 // Example: Expose it globally to the entire app (not recommended, prefer injection).
 Object.assign(window, { api })
 
+const app = createApp(Videos)
 app.mount('#app')
 
 // Example: Configure the fetch strategy.
@@ -23,5 +21,5 @@ Config.fetch = (...args) => {
 }
 
 // Example: Using the path helper, both path interpolation and request.
-console.log({ path: api.VideoClips.thumbnail.path({ id: 5, thumbnailId: 8 }) })
-api.VideoClips.latest().then((videos: any) => { console.log({ videos }) })
+console.log({ path: api.videoClips.thumbnail.path({ id: 5, thumbnailId: 8 }) })
+api.videoClips.latest().then((videos: any) => { console.log({ videos }) })
