+++ playground/vanilla/config/initializers/js_from_routes.rb
@@ -1,4 +1,6 @@
-# Example: Generate TypeScript files.
-JsFromRoutes.config do |config|
-  config.file_suffix = "Requests.ts"
+if Rails.env.development?
+  # Example: Generate TypeScript files.
+  JsFromRoutes.config do |config|
+    config.file_suffix = "Api.ts"
+  end
 end
